requirements.txt
python-telegram-bot>=20.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
pytz>=2023.3
pymongo>=4.0.0
